Treat false attribute values like null/undefined
Null and undefined attribute values are omitted, which is
convenient.

Having false attribute values be omitted would also be convenient.
Before this commit, a false value results in an actual attribute
added to the element, e.g.:

    const h = require('hyperscript')
    h('input', {type: 'text', name: null, required: false}).outerHTML

...which yields:

    '<input type="text" required="false">'

This commit makes false-valued attributes be omitted, so that the
above example would instead yield:

    '<input type="text">'